pyproject.toml
[tool.poetry]
name = "kursovaya-django-mailing"
version = "0.1.0"
description = ""
authors = ["denanmax <deshis93@gmail.com>"]
readme = "README.md"
packages = [{include = "kursovaya_django_mailing"}]

[tool.poetry.dependencies]
python = "^3.11"
django-scheduler = "^0.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
